# Enhance huge performance: reduce calculation time and memory allocation