--- a/go.mod
+++ b/go.mod
@@ -1,1 +1,3 @@
 module github.com/kelseyhightower/envconfig
+
+go 1.12