Before:
```
go test -v -bench=. -run=none \
         -benchtime=300000x -benchmem \
         -memprofile mem.out -cpuprofile cpu.out .
goos: linux
goarch: amd64
pkg: github.com/kelseyhightower/envconfig
BenchmarkGatherInfo-4             300000             39701 ns/op           17317 B/op        270 allocs/op
PASS
ok      github.com/kelseyhightower/envconfig    12.056s
```

After:
```
go test -v -bench=. -run=none \
        -benchtime=300000x -benchmem \
        -memprofile mem.out -cpuprofile cpu.out .
goos: linux
goarch: amd64
pkg: github.com/kelseyhightower/envconfig
BenchmarkGatherInfo-4             300000             31735 ns/op           16079 B/op        194 allocs/op
PASS
ok      github.com/kelseyhightower/envconfig    9.739s
```

Please note:

`ns/op`: 39701 -> 31735 **(-20%)**
`B/op`: 17317 -> 16079 **(-7.2%)**
`allocs/op`: 270 -> 194 **(-28.2%)**

----

Attachment with cpu and mem profile about difference:

(Github can't accept SVG of upload, so I compress as **GZ**)

[profile.tar.gz](https://github.com/kelseyhightower/envconfig/files/3513281/profile.tar.gz)